requirements.txt
APScheduler==3.7.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
flask-crontab==0.1.2
Flask-Script==2.0.6
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.1
pytz==2021.1
requests==2.25.1
six==1.15.0
torch==1.8.0+cpu
torchaudio==0.8.0
torchvision==0.9.0+cpu
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
Werkzeug==1.0.1
